requirements.txt
streamlit
tensorflow
opencv-python-headless
numpy
pillow
scikit-learn